pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dispatch-protocol"
version = "0.1.0"
description = "Operational delegation layer for autonomous AI agents"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "paramiko>=3.0",
    "httpx>=0.25",
    "rich>=13.0",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio", "ruff"]

[project.scripts]
dispatch = "src.cli:cli"

[tool.ruff]
line-length = 100
